+++ src/Data/Generics/Internal/Lens.hs
@@ -1,7 +1,9 @@
-{-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE GADTs         #-}
-{-# LANGUAGE Rank2Types    #-}
-{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE GADTs                  #-}
+{-# LANGUAGE LambdaCase             #-}
+{-# LANGUAGE Rank2Types             #-}
+{-# LANGUAGE TypeFamilies           #-}
+{-# LANGUAGE TypeFamilyDependencies #-}
+{-# LANGUAGE TypeOperators          #-}
 
 -----------------------------------------------------------------------------
 -- |
@@ -20,10 +22,11 @@ module Data.Generics.Internal.Lens where
 import Control.Applicative    (Const(..))
 import Data.Functor.Identity  (Identity(..))
 import Data.Monoid            (First (..))
-import Data.Profunctor        (Choice(right'), Profunctor(dimap))
+import Data.Profunctor        (Choice(..), Profunctor(..))
 import Data.Profunctor.Unsafe ((#.), (.#))
 import Data.Tagged
-import GHC.Generics           ((:*:)(..), (:+:)(..), Generic(..), M1(..), Rep)
+import GHC.Generics           ((:*:)(..), (:+:)(..), Generic(..), M1(..), K1(..), Rep)
+import Data.Coerce
 
 -- | Type alias for lens
 type Lens' s a
@@ -38,6 +41,9 @@ type Traversal' s a
 
 type Traversal s t a b
   = forall f. Applicative f => (a -> f b) -> s -> f t
+
+type LensLike f s t a b
+  = (a -> f b) -> s -> f t
 
 -- | Type alias for prism
 type Prism s t a b
@@ -85,20 +91,24 @@ infixr 8 #
 first :: Lens ((a :*: b) x) ((a' :*: b) x) (a x) (a' x)
 first f (a :*: b)
   = fmap (:*: b) (f a)
+{-# INLINE first #-}
 
 -- | Lens focusing on the second element of a product
 second :: Lens ((a :*: b) x) ((a :*: b') x) (b x) (b' x)
 second f (a :*: b)
   = fmap (a :*:) (f b)
+{-# INLINE second #-}
 
 left :: Prism ((a :+: c) x) ((b :+: c) x) (a x) (b x)
 left = prism L1 $ gsum Right (Left . R1)
+{-# INLINE left #-}
 
 leftP :: PrismP ((a :+: c) x) ((b :+: c) x) (a x) (b x)
 leftP = prismP L1 $ gsum Right (Left . R1)
 
 right :: Prism ((a :+: b) x) ((a :+: c) x) (b x) (c x)
 right = prism R1 $ gsum (Left . L1) Right
+{-# INLINE right #-}
 
 rightP :: PrismP ((a :+: b) x) ((a :+: c) x) (b x) (c x)
 rightP = prismP R1 $ gsum (Left . L1) Right
@@ -106,10 +116,7 @@ rightP = prismP R1 $ gsum (Left . L1) Right
 gsum :: (a x -> c) -> (b x -> c) -> ((a :+: b) x) -> c
 gsum f _ (L1 x) =  f x
 gsum _ g (R1 y) =  g y
-
-combine :: Lens' (s x) a -> Lens' (t x) a -> Lens' ((s :+: t) x) a
-combine sa _ f (L1 s) = fmap (\a -> L1 (set sa a s)) (f (s ^. sa))
-combine _ ta f (R1 t) = fmap (\a -> R1 (set ta a t)) (f (t ^. ta))
+{-# INLINE gsum #-}
 
 prism :: (b -> t) -> (s -> Either t a) -> Prism s t a b
 prism bt seta eta = dimap (\x -> plus pure id (seta x)) (either id (\x -> fmap bt x)) (right' eta)
@@ -123,10 +130,14 @@ plus :: (a -> b) -> (c -> d) -> Either a c -> Either b d
 plus f _ (Left x) = Left (f x)
 plus _ g (Right y) = Right (g y)
 
-
 -- | A type and its generic representation are isomorphic
 repIso :: (Generic a, Generic b) => Iso a b (Rep a x) (Rep b x)
-repIso = dimap from (fmap to)
+repIso = iso from to
+{-# INLINE repIso #-}
+
+repLens :: (Generic a, Generic b) => Lens a b (Rep a x) (Rep b x)
+repLens f s = repIso f s
+{-# INLINE repLens #-}
 
 -- | A type and its generic representation are isomorphic
 repIsoP :: (Generic a, Generic b) => IsoP a b (Rep a x) (Rep b x)
@@ -134,9 +145,9 @@ repIsoP = dimap from to
 
 
 -- | 'M1' is just a wrapper around `f p`
---mIso :: Iso' (M1 i c f p) (f p)
 mIso :: Iso (M1 i c f p) (M1 i c g p) (f p) (g p)
-mIso = dimap unM1 (fmap M1)
+mIso = iso unM1 M1
+{-# INLINE mIso #-}
 
 -- | 'M1' is just a wrapper around `f p`
 --mIso :: Iso' (M1 i c f p) (f p)
@@ -147,21 +158,24 @@ mIsoP = dimap unM1 M1
 -- these functions eta-expanded allows the optimiser to inline these functions.
 mLens :: Lens (M1 i c f p) (M1 i c g p) (f p) (g p)
 mLens f s = mIso f s
+{-# INLINE mLens #-}
 
-repLens :: (Generic a, Generic b) => Lens a b (Rep a x) (Rep b x)
-repLens f s = repIso f s
+kIso :: Iso (K1 r a p) (K1 r b p) a b
+kIso = iso unK1 K1
+{-# INLINE kIso #-}
 
-sumIso :: Iso' ((a :+: b) x) (Either (a x) (b x))
-sumIso = dimap f (fmap t)
-  where f (L1 x) = Left x
-        f (R1 x) = Right x
-        t (Left x) = L1 x
-        t (Right x) = R1 x
+sumIso :: Iso ((a :+: b) x) ((a' :+: b') x) (Either (a x) (b x)) (Either (a' x) (b' x))
+sumIso = iso back forth
+  where forth (Left l)  = (L1 l)
+        forth (Right r) = (R1 r)
+        back (L1 l) = (Left l)
+        back (R1 r) = (Right r)
+{-# INLINE sumIso #-}
 
-_Left :: Prism' (Either a c) a
+_Left :: Prism (Either a c) (Either b c) a b
 _Left = prism Left $ either Right (Left . Right)
 
-_Right :: Prism' (Either c a) a
+_Right :: Prism (Either c a) (Either c b) a b
 _Right = prism Right $ either (Left . Left) Right
 
 --------------------------------------------------------------------------------
@@ -181,72 +195,6 @@ proj fa = Coyoneda id fa
 ravel :: Functor f => ((a -> Coyoneda f b) -> s -> Coyoneda f t) -> (a -> f b) -> (s -> f t)
 ravel coy f s = inj $ coy (\a -> proj (f a)) s
 
---------------------------------------------------------------------------------
-
--- Pull dimaps to left and rights to right
-
-{-
-instance Choice p => Profunctor (MergeRight p) where
-  dimap f g (MergeRight (Dirty pab)) = MergeRight (DMap f g pab)
-  dimap f g (MergeRight (DMap f' g' pab)) = MergeRight (DMap (f' . f) (g . g') pab)
-  dimap f g (MergeRight (DRight pab))     =  MergeRight (DMap f g (right' pab))
-
-instance Choice p => Choice (MergeRight p) where
-  right' (MergeRight (Dirty pab)) = MergeRight (DRight pab)
-  right' (MergeRight (DMap f g pab)) = MergeRight (DMap (fmap f) (fmap g) (right' pab))
-  right' (MergeRight (DRight pab))   = MergeRight (DMap assoc' assoc (right' pab))
-  -}
-{-
-data PrismBoggle p a b where
-  DRight :: (p x y) -> PrismBoggle p (Either z x) (Either z y)
-  DMap :: (c -> a) -> (b -> d) -> (p a b) -> PrismBoggle p c d
-  Dirty :: p a b -> PrismBoggle p a b
-
-instance Choice p => Profunctor (PrismBoggle p) where
-  dimap f g (Dirty pab) = DMap f g pab
-  dimap f g (DMap f' g' pab) = DMap (f' . f) (g . g') pab
-  dimap f g (DRight pab)  =  DMap f g (right' pab)
-
-instance Choice p => Choice (PrismBoggle p) where
-  right' (Dirty pab) = DRight pab
-  right' (DMap f g pab) = DMap (fmap f) (fmap g) (right' pab)
-  right' (DRight pab)   = DMap assoc assoc' (right' pab)
-
-assoc :: Either a (Either b c) -> Either (Either a b) c
-assoc e = case e of
-            Left a -> Left (Left a)
-            Right ebc -> case ebc of
-                           Left b -> Left (Right b)
-                           Right c -> Right c
-
-assoc' :: Either (Either a b) c -> Either a (Either b c)
-assoc' e = case e of
-             Right c -> Right (Right c)
-             Left eab -> case eab of
-                           Left a -> Left a
-                           Right b -> Right (Left b)
-
-projPrism :: p a b -> PrismBoggle p a b
-projPrism pab = Dirty pab
-
---rightFuse :: Choice p => p a b -> p (Either c (Either d a)) (Either c (Either d b))
-rightFuse x = injPrism (injPrism (right' (right' (right' (projPrism (projPrism x))))))
-
-
-injPrism :: Choice p => PrismBoggle p a b -> p a b
-injPrism (Dirty pab) = pab
-injPrism (DMap f g pab) = dimap f g pab
-injPrism (DRight pab) = right' pab
--}
--- We don't need the full power of boggle, we just want to fuse together
--- two `fmaps`
-{-
-prismRavel ::
-                 (forall p . (MProfunctor p, MRight p) =>  p a b -> p s t)
-                 -> Prism s t a b
-prismRavel l pab = wrappedProfunctor (lowerFusionStack (ddimap id lowerBoggle (wrappedVLMProfunctor (ravelFusionStack l (WrappedVLMProfunctor (ddimap id (liftBoggle) (liftFusionStack (WrappedProfunctor pab))))))))
--}
-
 prismPRavel ::
                  (Market a b a b -> Market a b s t)
                  -> PrismP s t a b
@@ -259,16 +207,130 @@ prismRavel ::
 prismRavel l pab  = (prism2prismvl $ l idPrism) pab
 {-# INLINE prismRavel #-}
 
-------------------------------------------------------------------------------
--- Prism: Market
-------------------------------------------------------------------------------
+newtype AlongsideLeft f b a = AlongsideLeft { getAlongsideLeft :: f (a, b) }
+newtype AlongsideRight f b a = AlongsideRight { getAlongsideRight :: f (b, a) }
 
--- | This type is used internally by the 'Control.Lens.Prism.Prism' code to
--- provide efficient access to the two parts of a 'Prism'.
+(??) :: Functor f => f (a -> b) -> a -> f b
+fab ?? a = fmap ($ a) fab
+
+instance Functor f => Functor (AlongsideLeft f b) where
+  fmap f = AlongsideLeft . fmap (\(a, b) -> (f a, b)) . getAlongsideLeft
+
+instance Functor f => Functor (AlongsideRight f b) where
+  fmap f = AlongsideRight . fmap (\(a, b) -> (a, f b)) . getAlongsideRight
+
+alongside :: LensLike (AlongsideLeft f b') s  t  a  b
+          -> LensLike (AlongsideRight f t) s' t' a' b'
+          -> LensLike f (s, s') (t, t') (a, a') (b, b')
+alongside l1 l2 f (a1, a2)
+  = getAlongsideRight $ l2 ?? a2 $ \b2 -> AlongsideRight
+  $ getAlongsideLeft  $ l1 ?? a1 $ \b1 -> AlongsideLeft
+  $ f (b1,b2)
+{-# INLINE alongside #-}
+
+prodIso :: Iso ((a :*: b) x) ((a' :*: b') x) (a x, b x) (a' x, b' x)
+prodIso = iso (\(a :*: b) -> (a, b)) (\(a, b) -> (a :*: b))
+{-# INLINE prodIso #-}
+
+prodL :: Lens ((a :*: b) x) ((a' :*: b') x) (a x, b x) (a' x, b' x)
+prodL f s = prodIso f s
+{-# INLINE prodL #-}
+
+prodR :: Lens (a' x, b' x) (a x, b x) ((a' :*: b') x) ((a :*: b) x)
+prodR f s = fromIso prodIso f s
+{-# INLINE prodR #-}
+
+assoc3 :: Iso ((a, b), c) ((a', b'), c') (a, (b, c)) (a', (b', c'))
+assoc3 = iso (\((a, b), c) -> (a, (b, c))) (\(a, (b, c)) -> ((a, b), c))
+{-# INLINE assoc3 #-}
+
+assoc3L :: Lens ((a, b), c) ((a', b'), c') (a, (b, c)) (a', (b', c'))
+assoc3L f s = assoc3 f s
+{-# INLINE assoc3L #-}
+
+assoc3R :: Lens (a', (b', c')) (a, (b, c)) ((a', b'), c') ((a, b), c)
+assoc3R f s = fromIso assoc3 f s
+{-# INLINE assoc3R #-}
+
+choosing :: Lens s t a b -> Lens s' t' a b -> Lens (Either s s') (Either t t') a b
+choosing l _ f (Left a)   = Left <$> l f a
+choosing _ r f (Right a') = Right <$> r f a'
+{-# INLINE choosing #-}
+
+--------------------------------------------------------------------------------
+-- Iso stuff
+
+data Exchange a b s t = Exchange (s -> a) (b -> t)
+
+instance Functor (Exchange a b s) where
+  fmap f (Exchange sa bt) = Exchange sa (f . bt)
+  {-# INLINE fmap #-}
+
+instance Profunctor (Exchange a b) where
+  dimap f g (Exchange sa bt) = Exchange (sa . f) (g . bt)
+  {-# INLINE dimap #-}
+  lmap f (Exchange sa bt) = Exchange (sa . f) bt
+  {-# INLINE lmap #-}
+  rmap f (Exchange sa bt) = Exchange sa (f . bt)
+  {-# INLINE rmap #-}
+  ( #. ) _ = coerce
+  {-# INLINE ( #. ) #-}
+  ( .# ) p _ = coerce p
+  {-# INLINE ( .# ) #-}
+
+fromIso :: Iso s t a b -> Iso b a t s
+fromIso l = withIso l $ \ sa bt -> iso bt sa
+{-# INLINE fromIso #-}
+
+iso :: (s -> a) -> (b -> t) -> Iso s t a b
+iso sa bt = dimap sa (fmap bt)
+{-# INLINE iso #-}
+
+withIso :: Iso s t a b -> ((s -> a) -> (b -> t) -> r) -> r
+withIso ai k = case ai (Exchange id Identity) of
+  Exchange sa bt -> k sa (runIdentity #. bt)
+
+pairing :: Iso s t a b -> Iso s' t' a' b' -> Iso (s, s') (t, t') (a, a') (b, b')
+pairing f g = withIso f $ \ sa bt -> withIso g $ \s'a' b't' ->
+  iso (bmap sa s'a') (bmap bt b't')
+  where bmap f' g' (a, b) = (f' a, g' b)
+
+--------------------------------------------------------------------------------
+-- Prism stuff
+
+type APrism s t a b = Market a b a (Identity b) -> Market a b s (Identity t)
+
+-- without :: APrism s t a b
+--         -> APrism u v c d
+--         -> Prism (Either s u) (Either t v) (Either a c) (Either b d)
+-- without k =
+--   withPrism k         $ \bt seta k' ->
+--   withPrism k'        $ \dv uevc    ->
+--   prism (bimapE bt dv) $ \su ->
+--   case su of
+--     Left s  -> bimapE Left Left (seta s)
+--     Right u -> bimapE Right Right (uevc u)
+--   where bimapE :: (a -> b) -> (c -> d) -> Either a c -> Either b d
+--         bimapE f _ (Left a) = Left (f a)
+--         bimapE _ g (Right a) = Right (g a)
+--         {-# INLINE bimapE #-}
+
+without' :: APrism s t a b -> APrism s t c d -> Prism s t (Either a c) (Either b d)
+without' k =
+  withPrism k  $ \bt _ k' ->
+  withPrism k' $ \dt setc ->
+    prism (foldEither bt dt) $ \s -> fmap Right (setc s)
+  where foldEither _ g (Right r) = g r
+        foldEither f _ (Left l) = f l
+
+withPrism :: APrism s t a b -> ((b -> t) -> (s -> Either t a) -> r) -> r
+withPrism k f = case coerce (k (Market Identity Right)) of
+  Market bt seta -> f bt seta
+
+--------------------------------------------------------------------------------
+-- Market
+
 data Market a b s t = Market (b -> t) (s -> Either t a)
-
--- | @type 'Market'' a s t = 'Market' a a s t@
-type Market' a = Market a a
 
 instance Functor (Market a b s) where
   fmap f (Market bt seta) = Market (f . bt) (either (Left . f) Right . seta)
@@ -277,14 +339,22 @@ instance Functor (Market a b s) where
 instance Profunctor (Market a b) where
   dimap f g (Market bt seta) = Market (g . bt) (either (Left . g) Right . seta . f)
   {-# INLINE dimap #-}
+  lmap f (Market bt seta) = Market bt (seta . f)
+  {-# INLINE lmap #-}
+  rmap f (Market bt seta) = Market (f . bt) (either (Left . f) Right . seta)
+  {-# INLINE rmap #-}
+  ( #. ) _ = coerce
+  {-# INLINE ( #. ) #-}
+  ( .# ) p _ = coerce p
+  {-# INLINE ( .# ) #-}
 
 instance Choice (Market a b) where
-  right' (Market bt seta) = Market (Right . bt) $ \cs -> case cs of
-    Left c -> Left (Left c)
-    Right s -> case seta s of
-      Left t -> Left (Right t)
+  left' (Market bt seta) = Market (Left . bt) $ \sc -> case sc of
+    Left s -> case seta s of
+      Left t -> Left (Left t)
       Right a -> Right a
-  {-# INLINE right' #-}
+    Right c -> Left (Right c)
+  {-# INLINE left' #-}
 
 prism2prismp :: Market a b s t -> PrismP s t a b
 prism2prismp (Market bt seta) = prismP bt seta
@@ -381,6 +451,3 @@ instance Functor (Yoneda f) where
 instance Applicative f => Applicative (Yoneda f) where
   pure a = Yoneda (\f -> pure (f a))
   Yoneda m <*> Yoneda n = Yoneda (\f -> m (f .) <*> n id)
-
-
-