+++ src/Data/Generics/Product.hs
@@ -23,6 +23,7 @@ module Data.Generics.Product
   , module Data.Generics.Product.Subtype
   , module Data.Generics.Product.Typed
   , module Data.Generics.Product.Types
+  , module Data.Generics.Product.List
   ) where
 
 import Data.Generics.Product.Any
@@ -31,3 +32,4 @@ import Data.Generics.Product.Positions
 import Data.Generics.Product.Subtype
 import Data.Generics.Product.Typed
 import Data.Generics.Product.Types
+import Data.Generics.Product.List