+++ ChangeLog.md
@@ -1,5 +1,7 @@
 ## 0.5.1.0
-- Add `IsLabel` orhpan instance in `Data.Generics.Labels`
+- Infer input type from result type (#25)
+- Allow changing of multiple type parameters (#24)
+- Allow changing of type parameters that have kinds other than `*` (#23)
 - Fix error message in subtype lens
 
 ## 0.5.0.0